+++ src/GUI/AdminModule.java
@@ -8,15 +8,15 @@ package GUI;
 import Data.DataCSV;
 import Data.AnyToArrayList;
 import Domain.User;
-import GUI.CRUDadmin.CreateAdmin;
-import GUI.CRUDadmin.DeleteAdmin;
-import GUI.CRUDadmin.ListAdmins;
-import GUI.CRUDadmin.UpdateAdmin;
-import GUI.CRUDagent.CreateAgent;
-import GUI.CRUDagent.DeleteAgent;
-import GUI.CRUDagent.ListAgents;
-import GUI.CRUDagent.UpdateAgent;
-import GUI.RUDclient.ListClients;
+import GUI.CRUD.Admin.CreateAdmin;
+import GUI.CRUD.Admin.DeleteAdmin;
+import GUI.CRUD.Admin.ListAdmins;
+import GUI.CRUD.Admin.UpdateAdmin;
+import GUI.CRUD.Agent.CreateAgent;
+import GUI.CRUD.Agent.DeleteAgent;
+import GUI.CRUD.Agent.ListAgents;
+import GUI.CRUD.Agent.UpdateAgent;
+import GUI.CRUD.Client.ListClients;
 import Main.Algoritmos_Proyecto01_B16322_B31710_B67156;
 import Utilities.StringPath;
 import java.util.ArrayList;
