+++ src/GUI/CRUD/Admin/DeleteAdmin.form
