+++ src/GUI/CRUD/Agent/DeleteAgent.form
