+++ src/GUI/CRUD/Agent/DeleteAgent.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package GUI.CRUDagent;
+package GUI.CRUD.Agent;
 
 import Domain.User;
 import GUI.AdminModule;
