+++ src/GUI/CRUD/Client/ListClients.java
@@ -3,32 +3,31 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package GUI.CRUDadmin;
+package GUI.CRUD.Client;
 
-import Domain.User;
+import Domain.Client;
 import GUI.AdminModule;
-import GUI.AgentsModule;
 import Main.Algoritmos_Proyecto01_B16322_B31710_B67156;
 import java.util.ArrayList;
+import java.util.LinkedList;
 import javax.swing.table.DefaultTableModel;
 
 /**
  *
  * @author daniel
  */
-public class ListAdmins extends javax.swing.JFrame {
+public class ListClients extends javax.swing.JFrame {
 
     /**
      * Creates new form ListAgents
      */
     DefaultTableModel model;
-    ArrayList<User> admins;
-    
+    LinkedList<Client> clients;
 
-    public ListAdmins() {
+    public ListClients() {
         initComponents();
-        this.admins = Algoritmos_Proyecto01_B16322_B31710_B67156.ADMIN_LIST;
-        fillJtAgents(this.admins);
+        this.clients = Algoritmos_Proyecto01_B16322_B31710_B67156.CLIENT_LIST;
+        fillJtClient(this.clients);
     }
 
     /**
@@ -65,7 +64,7 @@ public class ListAdmins extends javax.swing.JFrame {
         ));
         jScrollPane1.setViewportView(jtAgents);
 
-        jButton1.setText("jButton1");
+        jButton1.setText("Reegresar");
         jButton1.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
                 jButton1ActionPerformed(evt);
@@ -83,9 +82,9 @@ public class ListAdmins extends javax.swing.JFrame {
                 .addGroup(jDesktopPane1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                     .addGroup(jDesktopPane1Layout.createSequentialGroup()
                         .addGap(20, 20, 20)
-                        .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 820, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 1102, javax.swing.GroupLayout.PREFERRED_SIZE))
                     .addGroup(jDesktopPane1Layout.createSequentialGroup()
-                        .addGap(352, 352, 352)
+                        .addGap(496, 496, 496)
                         .addComponent(jButton1)))
                 .addGap(20, 20, 20))
         );
@@ -94,9 +93,9 @@ public class ListAdmins extends javax.swing.JFrame {
             .addGroup(jDesktopPane1Layout.createSequentialGroup()
                 .addGap(30, 30, 30)
                 .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(30, 30, 30)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 35, Short.MAX_VALUE)
                 .addComponent(jButton1)
-                .addGap(30, 30, 30))
+                .addGap(25, 25, 25))
         );
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
@@ -106,7 +105,7 @@ public class ListAdmins extends javax.swing.JFrame {
             .addGroup(layout.createSequentialGroup()
                 .addGap(10, 10, 10)
                 .addComponent(jDesktopPane1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap(10, Short.MAX_VALUE))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -134,21 +133,27 @@ public class ListAdmins extends javax.swing.JFrame {
     private javax.swing.JScrollPane jScrollPane1;
     private javax.swing.JTable jtAgents;
     // End of variables declaration//GEN-END:variables
- public void fillJtAgents(ArrayList<User> agentList) {
+ public void fillJtClient(LinkedList<Client> clientList) {
 
         model = new DefaultTableModel();
 
+        model.addColumn("Id");
         model.addColumn("Nombre");
-        model.addColumn("Nombre de Usuario");
+        model.addColumn("Primer Apellido");
+        model.addColumn("Segundo Apellido");
         model.addColumn("E-Mail");
-        model.addColumn("Codigo");
+        model.addColumn("Teléfono");
+        model.addColumn("Provincia");
+        model.addColumn("Direccion Exacta");
 
         this.jtAgents.setModel(model);
 
-        if (!agentList.isEmpty()) {
+        if (!clientList.isEmpty()) {
 
-            for (User agent : admins) {
-                model.addRow(new Object[]{agent.getName(), agent.getUserName(), agent.getMail(), agent.getCode()});
+            for (Client client : clients) {
+                //mail, phoneNumber, province, exactAddress));
+                model.addRow(new Object[]{client.getId(), client.getName(), client.getLastNameA(), client.getLastNameB(), client.getMail(),
+                    client.getPhoneNumber(), client.getProvince(), client.getExactAddress()});
             }
 
         }
@@ -156,7 +161,6 @@ public class ListAdmins extends javax.swing.JFrame {
     }//Fin metodo que llena la tabla
 
     private void back() {
-        Algoritmos_Proyecto01_B16322_B31710_B67156.ADMIN_LIST = this.admins;
         this.dispose();
         AdminModule adminModule = new AdminModule();
         adminModule.setVisible(true);
