+++ src/GUI/CRUD/Client/UpdateClient.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package GUI.RUDclient;
+package GUI.CRUD.Client;
 
 import Domain.Client;
 import GUI.AdminModule;
